test(w5): lazy GPU probe を test_integration にも横展開 + guard を両ファイル化 (Refs #844)
